fix: unify binaries in one
