+++ src/client/method.rs
@@ -1,0 +1,7 @@
+use clap::ValueEnum;
+
+#[derive(ValueEnum, Clone, Debug)]
+pub enum Method {
+    Balance = 0,
+    Fund = 1,
+}
