+++ src/client/mod.rs
@@ -1,0 +1,2 @@
+pub mod balance;
+pub mod method;
